requirements.txt
psutil==5.9.8
python-vlc==3.0.20123
RPi.GPIO==0.7.1
